+++ other/tutorials/programming/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG for the programming workshop
 
+## 2019-08-03
+
+- using the [*tortoise*](https://q37.info/s/q4ghchsr) module.
+
 ## 2019-07-24
 
 - fixing issue concerning a bad separation of data between different sessions,